Disable search bar for play.oodle.ai
play.oodle.ai is where the users from Show HN
will get to explore a pre-created Oodle instance
in a read-only manner by auto-logging-in as a
hard-coded user.

Even a non-editor user in Grafana can set dark
mode for Grafana UI, which can make the Oodle UI
ugly. To make it difficult to accidentally do so,
and access any other similar settings, we disable
the search menu on play.oodle.ai.