Allow browser to indicate that a dashboard is updated (#116)
Ref D-5078

If the browser is certain that a dashboard is updated,
it can refresh the dashboard directly instead of relying
on websocket events which we have observed to be unreliable.

In this change, we allow an updater to include `::no_notification::`
in the dashboard update message, which disables server-sent
websocket notifications. Such an updater can emit an event in
the browser. We also add the listener for these events in this change.

One downside of this approach is that all the other instances of the
dashboard will never get notified, which is fine since anyway these
notifications do not have a guaranteed delivery and are just on a best
effort basis.